requirements.txt
chainlit~=2.5.5

autogen-agentchat~=0.5.3
autogen-ext[openai,web-surfer,file-surfer,magentic-one]~=0.5.3

markdown-pdf~=1.6
html2text~=2024.2.26
httpx~=0.27.2
bs4~=0.0.2
python-dotenv~=1.0.1

markitdown[all]
python-docx>=0.8.11
pypdf>=3.15.1
Markdown~=3.7
reportlab~=4.3.1
PyMuPDF>=1.25.3
weasyprint~=65.0
openai~=1.78.1
azure-storage-blob~=12.25.1
